Moved SkikoTestOwner to the ui-test module.
