+++ compose/ui/ui-test/src/skikoMain/kotlin/androidx/compose/ui/test/SkikoTestOwner.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.platform
+package androidx.compose.ui.test
 
 import androidx.compose.ui.InternalComposeUiApi
 import androidx.compose.ui.graphics.ImageBitmap
